Fixes #5 Vocabulary is not exposed in the properties metadata
