+++ spring-cloud-starter-stream-processor-tensorflow/pom.xml
@@ -1,67 +1,71 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-	<modelVersion>4.0.0</modelVersion>
+    <modelVersion>4.0.0</modelVersion>
 
-	<artifactId>spring-cloud-starter-stream-processor-tensorflow</artifactId>
-	<packaging>jar</packaging>
-	<name>spring-cloud-starter-stream-processor-tensorflow</name>
-	<description>Spring Cloud Stream processor-tensorflow core</description>
+    <artifactId>spring-cloud-starter-stream-processor-tensorflow</artifactId>
+    <packaging>jar</packaging>
+    <name>spring-cloud-starter-stream-processor-tensorflow</name>
+    <description>Spring Cloud Stream processor-tensorflow core</description>
 
-	<parent>
-		<groupId>org.springframework.cloud.stream.app</groupId>
-		<artifactId>tensorflow-app-starters-build</artifactId>
-		<version>1.0.0.BUILD-SNAPSHOT</version>
-	</parent>
+    <parent>
+        <groupId>org.springframework.cloud.stream.app</groupId>
+        <artifactId>tensorflow-app-starters-build</artifactId>
+        <version>1.0.0.BUILD-SNAPSHOT</version>
+    </parent>
 
-	<dependencies>
-		<dependency>
-			<groupId>org.tensorflow</groupId>
-			<artifactId>tensorflow</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-bus</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-starter-config</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud.stream.app</groupId>
-			<artifactId>tensorflow-app-starters-test-support</artifactId>
-		</dependency>
-	</dependencies>
+    <dependencies>
+        <dependency>
+            <groupId>org.tensorflow</groupId>
+            <artifactId>tensorflow</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-bus</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-config</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud.stream.app</groupId>
+            <artifactId>spring-cloud-starter-stream-common-tensorflow</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud.stream.app</groupId>
+            <artifactId>tensorflow-app-starters-test-support</artifactId>
+        </dependency>
+    </dependencies>
 
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.springframework.cloud</groupId>
-				<artifactId>spring-cloud-app-starter-doc-maven-plugin</artifactId>
-			</plugin>
-			<plugin>
-				<groupId>org.springframework.cloud.stream.app.plugin</groupId>
-				<artifactId>spring-cloud-stream-app-maven-plugin</artifactId>
-				<configuration>
-					<generatedProjectHome>${session.executionRootDirectory}/apps</generatedProjectHome>
-					<generatedProjectVersion>${project.version}</generatedProjectVersion>
-					<bom>
-						<name>scs-bom</name>
-						<groupId>org.springframework.cloud.stream.app</groupId>
-						<artifactId>tensorflow-app-dependencies</artifactId>
-						<version>${project.version}</version>
-					</bom>
-					<generatedApps>
-						<tensorflow-processor>
-							<extraTestConfigClass>
-								org.springframework.cloud.stream.app.test.tensorflow.TensorflowProcessorTestConfiguration.class
-							</extraTestConfigClass>
-						</tensorflow-processor>
-					</generatedApps>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.cloud</groupId>
+                <artifactId>spring-cloud-app-starter-doc-maven-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.springframework.cloud.stream.app.plugin</groupId>
+                <artifactId>spring-cloud-stream-app-maven-plugin</artifactId>
+                <configuration>
+                    <generatedProjectHome>${session.executionRootDirectory}/apps</generatedProjectHome>
+                    <generatedProjectVersion>${project.version}</generatedProjectVersion>
+                    <bom>
+                        <name>scs-bom</name>
+                        <groupId>org.springframework.cloud.stream.app</groupId>
+                        <artifactId>tensorflow-app-dependencies</artifactId>
+                        <version>${project.version}</version>
+                    </bom>
+                    <generatedApps>
+                        <tensorflow-processor>
+                            <extraTestConfigClass>
+                                org.springframework.cloud.stream.app.test.tensorflow.TensorflowProcessorTestConfiguration.class
+                            </extraTestConfigClass>
+                        </tensorflow-processor>
+                    </generatedApps>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
