+++ spring-cloud-starter-stream-processor-tensorflow/src/main/java/org/springframework/cloud/stream/app/tensorflow/processor/TensorflowProcessorConfiguration.java
@@ -16,30 +16,21 @@
 
 package org.springframework.cloud.stream.app.tensorflow.processor;
 
-import java.io.IOException;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.springframework.context.annotation.Import;
 import org.tensorflow.Tensor;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.expression.EvaluationContext;
-import org.springframework.integration.annotation.ServiceActivator;
-import org.springframework.integration.context.IntegrationContextUtils;
-import org.springframework.integration.support.MessageBuilder;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.converter.MessageConversionException;
 import org.springframework.tuple.Tuple;
-import org.springframework.tuple.TupleBuilder;
 
 /**
  * A processor that evaluates a machine learning model stored in TensorFlow's ProtoBuf format.
@@ -66,120 +57,11 @@ import org.springframework.tuple.TupleBuilder;
  * @author Artem Bilan
  */
 @EnableBinding(Processor.class)
-@EnableConfigurationProperties(TensorflowProcessorProperties.class)
-public class TensorflowProcessorConfiguration implements AutoCloseable {
+@EnableConfigurationProperties(TensorflowCommonProcessorProperties.class)
+@Import(TensorflowCommonProcessorConfiguration.class)
+public class TensorflowProcessorConfiguration {
 
 	public static final String ORIGINAL_INPUT_DATA = "original.input.data";
 
 	private static final Log logger = LogFactory.getLog(TensorflowProcessorConfiguration.class);
-
-	@Autowired
-	@Qualifier(IntegrationContextUtils.INTEGRATION_EVALUATION_CONTEXT_BEAN_NAME)
-	private EvaluationContext evaluationContext;
-
-	@Autowired
-	private TensorflowProcessorProperties properties;
-
-	@Autowired
-	@Qualifier("tensorflowInputConverter")
-	private TensorflowInputConverter tensorflowInputConverter;
-
-	@Autowired
-	@Qualifier("tensorflowOutputConverter")
-	private TensorflowOutputConverter tensorflowOutputConverter;
-
-	@Autowired
-	private TensorFlowService tensorFlowService;
-
-	@ServiceActivator(inputChannel = Processor.INPUT, outputChannel = Processor.OUTPUT)
-	public Object evaluate(Message<?> input) {
-
-		Object inputData =
-				this.properties.getExpression() == null
-						? input.getPayload()
-						: this.properties.getExpression().getValue(this.evaluationContext, input, Object.class);
-
-		// The processorContext allows to convey metadata from the Input to Output converter.
-		Map<String, Object> processorContext = new ConcurrentHashMap<>();
-
-		Map<String, Object> inputDataMap = this.tensorflowInputConverter.convert(inputData, processorContext);
-
-		Tensor outputTensor = this.tensorFlowService.evaluate(inputDataMap, this.properties.getModelFetch(),
-				this.properties.getModelFetchIndex());
-
-		Object outputData = tensorflowOutputConverter.convert(outputTensor, processorContext);
-
-		switch (this.properties.getMode()) {
-
-			case tuple:
-				TupleBuilder outTupleBuilder = TupleBuilder.tuple().put(properties.getOutputName(), outputData);
-
-				Object payload = input.getPayload();
-
-				if (payload instanceof Tuple && ((Tuple) payload).hasFieldName(ORIGINAL_INPUT_DATA)) {
-					// If the payload is already a tuple that contains ORIGINAL_INPUT_DATA entry then copy the
-					// content of the input tuple in the new tuple to be returned.
-					outTupleBuilder.putAll((Tuple) payload);
-				}
-				else {
-					// This is a new tuple so preserve the input data.
-					outTupleBuilder.put(ORIGINAL_INPUT_DATA, payload);
-				}
-
-				return outTupleBuilder.build();
-
-			case header:
-				return MessageBuilder
-						.withPayload(input.getPayload())
-						.setHeader(this.properties.getOutputName(), outputData);
-
-			default:
-				return outputData;
-		}
-
-	}
-
-	@Bean
-	@RefreshScope
-	public TensorFlowService tensorFlowService() throws IOException {
-		return new TensorFlowService(this.properties.getModel());
-	}
-
-	@Bean
-	@ConditionalOnMissingBean(name = "tensorflowOutputConverter")
-	public TensorflowOutputConverter tensorflowOutputConverter() {
-		// Default implementations serializes the Tensor into Tuple
-		return new TensorflowOutputConverter<Tuple>() {
-
-			@Override
-			public Tuple convert(Tensor tensor, Map<String, Object> processorContext) {
-				return TensorTupleConverter.toTuple(tensor);
-			}
-		};
-	}
-
-	@Bean
-	@ConditionalOnMissingBean(name = "tensorflowInputConverter")
-	@SuppressWarnings("unchecked")
-	public TensorflowInputConverter tensorflowInputConverter() {
-		return new TensorflowInputConverter() {
-
-			@Override
-			public Map<String, Object> convert(Object input, Map<String, Object> processorContext) {
-
-				if (input instanceof Map) {
-					return (Map<String, Object>) input;
-				}
-
-				throw new MessageConversionException("Unsupported input format: " + input);
-			}
-		};
-	}
-
-	@Override
-	public void close() throws Exception {
-		logger.info("Close TensorflowProcessorConfiguration");
-		tensorFlowService.close();
-	}
-
 }
