+++ spring-cloud-starter-stream-processor-tensorflow/src/main/resources/META-INF/spring-configuration-metadata-whitelist.properties
@@ -1,1 +1,1 @@
-configuration-properties.classes=org.springframework.cloud.stream.app.tensorflow.processor.TensorflowProcessorProperties
+configuration-properties.classes=org.springframework.cloud.stream.app.tensorflow.processor.TensorflowCommonProcessorProperties
