+++ spring-cloud-starter-stream-processor-twitter-sentiment/pom.xml
@@ -20,8 +20,8 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud.stream.app</groupId>
-			<artifactId>spring-cloud-starter-stream-processor-tensorflow</artifactId>
-		</dependency>		
+			<artifactId>spring-cloud-starter-stream-common-tensorflow</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
