+++ spring-cloud-starter-stream-processor-twitter-sentiment/src/main/resources/META-INF/spring-configuration-metadata-whitelist.properties
@@ -1,2 +1,2 @@
 configuration-properties.classes=org.springframework.cloud.stream.app.twitter.sentiment.processor.TwitterSentimentProcessorProperties,\
-   org.springframework.cloud.stream.app.tensorflow.processor.TensorflowProcessorProperties
+   org.springframework.cloud.stream.app.tensorflow.processor.TensorflowCommonProcessorProperties
