+++ tensorflow-app-dependencies/pom.xml
@@ -30,6 +30,11 @@
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.cloud.stream.app</groupId>
+				<artifactId>spring-cloud-starter-stream-common-tensorflow</artifactId>
+				<version>1.0.0.BUILD-SNAPSHOT</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud.stream.app</groupId>
 				<artifactId>spring-cloud-starter-stream-processor-tensorflow</artifactId>
 				<version>1.0.0.BUILD-SNAPSHOT</version>
 			</dependency>
